refactor: [Frontend] changed folder structure
