+++ frontend/src/global_functions/Tools.tsx
