+++ frontend/src/models/ScenarioTypes.tsx
