+++ frontend/src/pages/Main/MainPage.tsx
@@ -1,5 +1,5 @@
-import {UserInfos} from "../../Types.tsx";
-import ScenarioList from "../scenario/ScenarioList.tsx";
+import {UserInfos} from "../../models/BaseTypes.tsx";
+import ScenariosPage from "../Scenarios/ScenariosPage.tsx";
 
 type Props = {
     user?: UserInfos
@@ -14,5 +14,5 @@ export default function MainPage( props: Readonly<Props> ) {
     if (!props.user.isUser && !props.user.isAdmin)
         return <><br/>You are now logged in, but should wait until an administrator grants you access to the app.</>
 
-    return <ScenarioList user={props.user}/>
+    return <ScenariosPage user={props.user}/>
 }
