+++ frontend/src/pages/Main/components/DarkModeSwitch.Functions.tsx
