+++ frontend/src/pages/Main/components/SidePanel.css
