+++ frontend/src/pages/Main/components/SidePanel.tsx
@@ -1,6 +1,6 @@
 import "./SidePanel.css"
 import {ReactNode, useState} from "react";
-import {SHOW_RENDERING_HINTS} from "../../Types.tsx";
+import {SHOW_RENDERING_HINTS} from "../../../models/BaseTypes.tsx";
 
 type Props = {
     children: ReactNode[]
