+++ frontend/src/pages/NewTestRun/components/VariablesEdit.tsx
@@ -1,8 +1,8 @@
 import StringListInput from "./StringListInput.tsx";
 import {useEffect, useState} from "react";
 import styled from "styled-components";
-import {SHOW_RENDERING_HINTS} from "../../../Types.tsx";
-import {VariablesChangeMethod} from "../Types.tsx";
+import {SHOW_RENDERING_HINTS} from "../../../models/BaseTypes.tsx";
+import {VariablesChangeMethod} from "../../../models/TestRunTypes.tsx";
 
 const SimpleCard = styled.div`
   border: 1px solid var(--border-color, #707070);
