+++ frontend/src/pages/SimpleChat/SimpleChatPage.tsx
@@ -1,7 +1,7 @@
-import './SimpleChatView.css';
+import './SimpleChatPage.css';
 import {ChangeEvent, FormEvent, useState} from "react";
 import axios from "axios";
-import {Answer, DEBUG, Prompt, SHOW_RENDERING_HINTS} from "../Types.tsx";
+import {Answer, DEBUG, Prompt, SHOW_RENDERING_HINTS} from "../../models/BaseTypes.tsx";
 
 export default function SimpleChatPage() {
     const [ prompt, setPrompt ] = useState<Prompt>({ prompt:"" });
